pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "A Family Tree App"
authors = [{ name = "Elshaday97", email = "yirga.elshaday@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "neo4j (>=6.1.0,<7.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "phonenumbers (>=9.0.24,<10.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi-users[oauth,sqlalchemy] (>=15.0.4,<16.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
]

[tool.potery]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
